Move assets folder to sdkproviderutils
For the mediatee SDK to be able to load
a webview with content from assets, the
mediatee SDK should have access to assets
directory. We should move the assets folder
to a common directory where the mediatee
SDK can also access it and load an ad with
content from assets.

Bug: 346738801
Test: Ui Lib Manual Test App
Change-Id: I8bfce7e764fd35aa2474c6689b2ecd24898a2d9c