Update the existing use cases when add new use cases.
When use cases are bound, the implementation options of the already
bound use cases such as stream_use_case should be updated based on the
new use cases.

This resolves the issue that the capture session failed to configure when
a Preview and a VideoCapture are bound separately due to unsupported
stream_use_case combination.

Bug: b/342977497
Test: ./gradlew camera:camera-core:build
Change-Id: Ia256fd2eb776ebdab5dcd53c4a20e15f7328f2d2