Add utility class for viewability measurement
Adds a utility class to the UI library manual test app to allow
either the client or provider process to start measurement for
a SandboxedUiAdapter, optionally drawing rectangles representing
the recorded on screen geometry. This ensures that the
viewability rectangles can be drawn during app owned mediation.
Otherwise, this would fail as the provider process would try to
draw an overlay on a view that it does not own.

Test: manual test app: turn on app owned mediation and draw
      viewability toggles
Bug: 348610737
Change-Id: I1a974e4e7fc7e8438905cf3308f05265b76ebaf7