Move transformToScreen to android-only extension
`LayoutCoordinates.transformToScreen` was added to provide matrix to Android platform API, but it brings a requirement to implement it for all platforms. However some platforms considers transform matrices as private representation so it cannot be always properly supported.

This change moves it to android-only extension.

RelNote: Move `LayoutCoordinates.transformToScreen` to android-only extension
Test: exiting unit tests and demos
Change-Id: Ic540aa1b4bda9d10a999bef2430cdefb52546571