Fixes overscroll not propagating fling velocity used to relax a stretch
Previously (before aosp/3022171) we were incorrectly stretching, but not relaxing, when applying a fling before the scrollable itself has a chance to fling. Now that is fixed, this uncovers a separate issue where if we are stretched, and we are flinging to relax the stretch, all the velocity will be consumed and the list will not be able to scroll. There is also a separate issue where we always release existing stretches when we move to relax (releaseOppositeOverscroll), even though we still should be consuming more delta before we relax.

This cl fixes this behavior such that:

- If a fling does not have enough velocity to fully relax the stretch, we consume all the velocity and the stretch will recede on its own
- If a fling has enough velocity to relax the stretch, instead of consuming velocity, we consume the delta as part of applyToScroll caused by the fling. Any leftover delta after we have fully relaxed will continue to be used to scroll.
- We only release opposite overscroll if there is delta left for scrolling (i.e this will only work for active glow effects, not stretch effects), so stretches won't start to relax as soon as you move in the opposite direction - they will remain until they are fully relaxed with a drag / fling.

Also cleans up / fixes some incorrect invalidation logic.

Test: OverscrollTest
Test: StretchOverscrollIntegrationTest
Change-Id: Ib0880f0e326e491b7927f6ec780e6cf867017dee