Refactor tests to support the new InCallService Extensions API
1) Rename MockInCallServiceDelegate ->TestInCallService and simplify
the implementation to only use one impl of InCallServiceCompat
2) Move from a static impl of TestInCallService to a scoped model
where BaseTelecomTest allows users to call `usingIcs` and bind to
the ICS for a specific test. This has the advantage of ensuring that
we clean up any calls that were left behind.
3) Remove CallCompat, CapabilityExchange, and replace ParticipantClientExtension
with the new version of it. Clean up the tests to reflect this change.
4) Removed the InCallServiceCompatTest because it was flaky. In a follow up
CL, we will need to figure out how to better unit test these clasees by
creating interfaces.

Relnote: Implements ICS Extensions API for voip app actions
Bug: 338651142
Test: core-telecom tests on V/U/S

Change-Id: I1274ca5cc898b0bd20ad90a1b97ee0ae8c90861b