Fix stack trace of exception thrown in runOnUiThread
The original exception caught on the UI thread was rethrown without
adding any information about the call site. A suppressed exception
created on the call site is added, which shows the method call that
switched to the UI thread.

Fix: 350050393
Test: runOnUiThread_exception_rethrownWithCallersStackTrace()
Relnote: "Exceptions thrown during an action passed to
`ComposeUiTest.runOnUiThread()` or `ComposeTestRule.runOnUiThread()` now
have a suppressed exception added that was created on the caller's side,
to follow the stacktrace across threads."

Change-Id: I17421876b3f26bec7c064a5f0874bfdcb0782aa0