Send ancestor events for LiveRegion announcements
Prior to this CL, if an ancestor and its descendant both had layout changes, only the descendant would send accessibility events (this was done in aosp/2729277 to minimize number of events sent by Compose).

In the case of a liveRegion activated by a button, if the ancestor value change was triggered by a descendant button being clicked, only the `TYPE_VIEW_CLICKED` event from the descendant was being clicked, even though the ancestor value was also being changed. Without alerting TalkBack that the ancestor value was being changed, LiveRegions were not working properly.

With this fix, `CONTENT_CHANGE_TYPE_SUBTREE` is now sent in addition to `TYPE_VIEW_CLICKED`, fixing the scenario where liveRegion announcements were not being recognized by TalkBack if the change and its trigger were on ancestor/descendant components.

Fixes: 348590026
Test: sendClickedAndChangeSubtree_whenDescendantClicked() and verified with local repro provided in bug
Change-Id: I308d5122947ecc3afbf0dbeb0d96f977d655be8a