Releases stretch overscroll if scroll consumes
If stretch overscroll is showing because a list is scrolled to the end and new items are added, the next scroll event will start to scroll the list but stretch overscroll will stay stretched. This CL changes it so that if scroll was consumed while a stretch is active, the stretch will be released so that the list can scroll without being stretched. This isn't needed for glow, since glow will fade out on its own if it isn't continously pulled, whereas stretch will stay until released.

Fixes: b/262253616
Test: StretchOverscrollIntegrationTest
Change-Id: Icccab49ab5b08b8b61be9bfb0bad636a9fc87e3f