Apply spans from setComposingText to text layout in BTF2
IMEs can attach certain visual clues to composing region for users to easily understand the current state of composition of a word or a phrase. This feature is heavily used in Japanese keyboards.

TextFields in Compose have always ignored the possibly existing non-text parts of `CharSequence` in `InputConnection`. This change basically adds the ability to convert some subset of known Spans to their Compose counterparts in `SpanStyle` or `ParagraphStyle` to be rendered by the new BasicTextField.

Test: gradle :compose:foundation:foundation:cAT
Bug: 335559344
Change-Id: I9ff0e101961cb335e7aa8ebc35f90e28dd52f2cc