Publish LazyLayoutAnimateScrollScope and factories for common layouts.
In this CL we're making LazyLayoutAnimateScrollScope public to allow customization of animated scroll in LazyLayouts. We're also providing factories for creating LazyLayoutAnimateScrollScopes for Lists, Grids and Pager. Doc: go/lazy-layouts-next-scroll-animations

Test: Existing tests.
Bug: 218542111
Relnote: Making LazyLayoutAnimateScrollScope public to allow customization of animated scroll in LazyLayouts. We're also providing factories for creating LazyLayoutAnimateScrollScopes for Lists, Grids and Pager.
Change-Id: I796e2004d20ea1b529e4ecb365b34013a0a5eb8f