Make class bound go before interface bounds
In Java a class type parameter bound should go before interface bounds.
Here we make an extra sorting in KspTypeParameterElement.bounds to ensure that. Though this ordering could be different from the declaration order in Kotlin, it makes codegen easier and also makes the behavior a bit more consistent with KAPT.

Test: XTypeParameterElementTest
Bug: https://github.com/google/dagger/issues/4345
Change-Id: Ifd766607dab08235df2565dc1b964d4293289af9