Preserve the test APKs after a test run
Added a android.injected.androidTest.leaveApksInstalledAfterRun=true to
the gradle.properties to prevent the APKs and any test outputs from
being deleted after a test run.
The output files for those runs also contains screenshot files that can
be inspected before a test is uploaded to gerrit, so it's useful to have
those stick around.

Test: Manual - Ran a screenshot tests and checked that the generated
files are accessible through the Device Explorer

Change-Id: Ib44a4dff8d77e23f9ad1b0f25a162dccb5da1e66