Implement alwaysMinimizeLabel for text fields
Equivalent to the boolean opposite of expandedHintEnabled
on Views. This property allows forcing the label to be
minimized even when the text field is unfocused. If the label
is minimized, this allows the placeholder and prefix/suffix
text to display even if the text field is unfocused.

Fixes: b/307794529
Test: added unit and screenshot tests
Relnote: "Added `alwaysMinimizeLabel` parameter to
  text fields. Setting this value to `true` allows
  the UI pattern of displaying a label and placeholder
  in a text field at the same time even when the field
  is unfocused."
Change-Id: I8d0ba68d4b323d90b5b3c1de1b319aeb2ef8ac83