Rehash in place when dropping deleted entries
The previous implementation only cleaned up the metadata to drop
deleted entries when the map/set was getting full. This caused one
of the invariants of the data structure to be violated, leading
to corrupted queries.

When we want to drop the deleted entries, we must re-probe the
data structure to move the entries in the proper group based on
their hash. Failure to do so means an entry could remain in a
spot where it should only land in the event of a collision.

This change reworks how we clone the metadata at the end of the
array to be slightly more efficient. We could save further
instructions but it could be at the expensve of the data cache
so we're going to leave this as an exercise for our future
selves.

Also cleanup LruCache (set the upper bound of the maxSize
range to be valid for Int storage, pre-size the map used
in snapshot()).

Fixes: 352560465
Test: ScatterMapTest, ScatterSetTest and primitive variants
Relnote: """Fixed a corruption that could happen in ScatterMap,
ScatterSet, and their primitive variants."""

Change-Id: I38a4afee75595958594310a0ad14cca486997163