Recycle connection if they are undelivered.
Closing SQLite connections on the onUndeliveredElement of the Pool's Channel caused for the Channel to not be refilled with a new connections, ultimately leading to timeouts during connection acquisition. This CL updates the onUndeliveredElement to simply recycle the unused connection, refilling the pool and letting new waiters acquire the unused connection.

Bug: 347737870
Test: BaseConnectionPoolTest
Change-Id: Ic5cac60a1855bccea99f476cf6236ea397004710