Refactor TestContext and TestOwner
Mostly this turns public API marked as @InternalTestApi into internal
members, which is much more appropriate. It also makes a better
separation between the two, putting more emphasize on TestContext as a
state holder and TestOwner as a link to production layers (i.e.
retrieval of the semantics tree).

Bug: N/A
Test: bOS
Relnote: "Made TestOwner, which was annotated as @InternalTestApi,
internal."

Change-Id: I92138974bf077f4aa7ba1d50b71b8899c91c76bb