Fix SharedTransitionScope not drawing items in certain conditions
Originally, there was nothing within SharedTransitionScope that could explicitly cause an invalidation, so in some cases, after at least a transition with different content has ocurred, there may be nothing (as in, an obvservable state) that would cause the SharedTransitionScope to draw.

So nothing would get drawn on the overlay during transition.

Although the measure/layout and corresponding animation would still happen within the layout hierarchy.

This fixes it by making the rendering target list (for the overlay) a MutableStateList. Which is updated as the content changes for a transition.

Bug: 347520198
Test: testSharedTransitionScopeIsInvalidated
Change-Id: Id65abf8cc50fc08e246c0a85581df32be1983ecc