Reduce synchronized blocks in SystemJobService
All 3 methods that interact with 'mJobParameters', onStartJob(), onStopJob() and onExecuted() are invoked in the main thread, so there is no need to synchronize interactions with the map. Similarly, StartStopTokens interactions are also done in the main thread, so we can use a version that does not use synchronized blocks.

This changes are done in an attempt to reduce main thread time during onStartJob() and onStopJob() since with Android 14, they can cause an ANR if WorkManager takes too long to complete the method.

Bug: 353295716
Test: Existing
Change-Id: I1319ded3b099ae614eafdfbeb5314f9aa31e724b