Ignore unsupported capability
When decoding the capabilities from the Work database, ignore invalid capabilities by catching the IllegalArgumentException thrown by the framework. This is an attempt to mitigate b/351180465 which can happen on a newer OS adding new default capabilities but the device downgrading and not wiping the app's data and thus using the database persist capabilities of the newer OS version.

Bug: 351180465
Test: WorkConstrainsDaoTest
Change-Id: I063e1995fa5b40cdb6153db3e13319ed26670990