Add a `shouldExecute` API to use instead of calling `skipping`.
Introduce a new compiler API that takes the place of calling
`skipping` directly as it give the runtime more control on
when the body of a function executes.

This is a step in supporting pausable composition that allows
the compiler code generation changes to be tested without
requiring the runtime to fully support pausable composition.

BUG: 328817808
Test: existing tests, ./gradlew :compose:r:r:tDUT
Relnote: N/A - Compiler API change
Change-Id: I36ea4b29e7f545265c175e1aa09c67b57ff7ed77