StartDestination route to use default arg values
In aosp/2978197, we added a fix so navigation would parse and store arg values passed through the startDestination route. However, this behavior change is not ideal for existing implementations that rely on the ignored args to fallback to default values.

This CL aims to adjust the behavior so that we support both default arg and passed in arg for startDestination.

If the startDestination route is identical to the startDestination.route (all args are placeholders), we will fallback to default value for all the args. If at least one arg placeholder has been replaced with an actual value, then the arg and placeholders will be parse and handled as actual arg values.

As an example:

Given this startDestination route "example.com/{firstName}/{lastName}"

1. if startDest = "example.com/Tom/{lastName}", bundle would contain [Tom, {lastName}]
2. if startDest = "example.com/{firstName}/{lastName}", bundle would contain the default values

Note that in case.1, placeholder values i.e. {lastName} may not be parsed successfully. If {lastName} were of NavType.IntType, the parse would fail and this value would not be added to the bundle.

Test: ./gradlew navigation:navigation-runtime:cC
Bug: 354046047
Bug: 342016424
Relnote: "StartDestination will use default arg values if the startDestination route passed to the NavGraph is exactly equal to the startDestination.route"
Change-Id: I137626c99c62ec57db4a44448dd7cdf61209f264