Experiment: Enable R8 in foundation-library and compose-runtime benchmarks
Bug: 354264743

Benchmarks appear to bounce around when something in the library
changes, regardless of whether it should affect the benchmark.
For example, adding a new collection could cause a benchmark
to increase or decrease, even though the collection hasn't
been used anywhere.

This CL set minifyEnabled to true for compose foundation-library
and compose runtime benchmarks, two libraries that have seen
recent performance variability. This should be considered an
experiment that may be removed in the future or expanded to
other libraries, depending on whether this helps performance
alerts.

Test: ran perf tests locally
Change-Id: I5f0b06631b01e9491bffc5c005b15e7153aa8787