Fix CameraController resolution configuration issue.
If ResolutionSelector and OutputSize are both set to CameraController,
the configuration will conflicts. A ResolutionSelector will also be set
when the viewport is configured without taking account of whether
OutputSize is set or not.

This CL makes ResolutionSelector takes precedence to OutputSize when
used together. And only override the aspect ratio when they are both not
set.

Bug: b/352737959
Test: ./gradlew camera:camera-view:build
Change-Id: I5d99600bba9f694feb37ad65a81f423ea6f89317