Don't return collapsedSemantics on unattached or deactivated nodes
This patch ensures that if a node is deactivated but not attached then `collapsedSemantics` are not returned (and `applySemantics` is not called). This addresses the exception thrown when `applySemantics` is called on an unattached node.

In returning null for deactivated nodes, the SemanticsNodeInteraction.assertIsDeactivated assert needs to be refactored. We need to surface deactivated nodes but cannot do so with testTags because testTags come from semantics config, which comes from modifier nodes. This patch rewrites all foundation tests using the `assertIsDeactivated` assert by instead saving the semantics ids of relevant nodes, then verifying that those nodes all are deactivated by checking `layoutInfo.isDeactivated`.

Test: manually verified Material catalog does not crash
Fixes: 353595612
Fixes: 353538515
RelNote: "New `semanticsId()` test API introduced."
Change-Id: Iee26db6343fbaa3689546256eac2bb0d9725adea