Create navigation-lint-common artifact
Artifact to house lint utils for Navigation.

Lint checks for safe args in common/runtime/compose artifacts share common logic. Currently the common logic is placed within `navigation-common-lint` which is the lint module for `navigation-common`. However, runtime/compose lints that are bundled with common-lint leads to `Found duplicate NavigationCommonIssueRegistry` issues.

So we need to host the shared utils in a separate lint module that does not contain any IssueRegistry.

Test: TH
Bug: 342016424
Change-Id: I3e8611e056005b5f0fe650a295661b18ebfb864a