Marking `invisibleToUser()` as Experimental
This API was experimental before the experimental policy change in aosp/1541325.

After receiving API feedback about renaming it in aosp/3177464, for now we will keep this API as experimental while we discuss with API council and the UI Understanding team about 1) what to name the API, 2) what best practices we will recommend for developers in the future about when to use this API (particularly as opposed to `clearAndSetSemantics()`.

Doc for this discussion and notes at go/stabilize-invisibletouser

Test: N/A
RelNote: "Marking `invisibleToUser()` as Experimental."
Change-Id: I8108fc89f527225767544951c57837ff8901d310