Synchronize interop internals between platforms (#1462)
Refactor of interop internals, the behavior is not changed.

Key changes:
- Proper re-use of common `InteropView` typealias on Desktop
- Introduce internal `InteropViewGroup` typealias to reuse
wrapper/container logic without generic parameters
- Re-use of common `InteropViewFactoryHolder` typealias. There were
holders on all platforms but it wasn't united in any way
- Re-use of interop `PointerInputFilter` that used for hit testing and
input re-dispatching
- Remove generic parameter from `InteropContainer`
- Move `LocalInteropContainer` to skiko source set, clean up APIs so no
need for platform specifics anymore here
- Move copies of `Updater` to skiko sourceset

Note that this is not the final shape, but the first step in a direction
of reusing more code and uniting the approaches between platforms
(including Android). I've added a few TODOs in places that should be
changed.