Fixes clickable and combinedClickable not emitting PressInteraction.Cancel when they lose focus, while tracking a key event
Previously if a down key was being handled and focus was lost (such as by tabbing away), a Cancel was never emitted, so ripples and other effects observing this would stay indefinitely pressed.

This also properly disposes existing long key presses in combinedClickable, so a long click won't be triggered if focus is lost before enough time elapses.

Also adds ClickableParameterizedKeyInputTest to match the one for combinedClickable

Fixes: b/352093462
Test: Clickable / CombinedClickable* tests
Change-Id: If7b62c1131cd40ead928f7801a45e72d73c8574c