Remove @ExperimentalCarApi tags from androidx.car.app.messaging.model classes
APIs have been fully functional for a while, and are exercised in
production by multiple internal Google apps. We definitely want to enforce compat.
Also, it makes sense at this point to allow 3p apps to use the APIs, as they are
fairly mature.

Bug: 355109181
Test: ./gradlew :car:app:app:test
Relnote: "Mark messaging APIs as non-experimental"
Change-Id: I0b0700c03fe032f433db8e3de4bb0b4c8867948d