Fixes the crash issue that occurs after scrolling to page beyond the previous PDFs total pages.
Root cause: On cancelling the previous loader's tasks, an error is
reported in the GetDimensionsTask which calls the pageBroken() callback. Here it tries to access a page which is more than the number of pages and ArrayOutOfBoundException is thrown. This is fixed by adding a if check to check if the page num is less the number of pages of pdf.

Bug: 356188262
Test: manually tested
Change-Id: Ib0e04d8acab87fc5fe940c1208ace5426f8cd351