Fix extensions ReleaseApkTest failure for devices missing essential impl files
Some devices might remove ImageCaptureExtenderImpl or PreviewExtenderImpl files. Need to skip the test for those extension modes by calling checkTestlibRunning() to ensure that it is not running in testlib implementation.

Also, return empty list if the testing device is not in the test lab. By verifying on go/forrest, it looks like that this can avoid the EXPECTED_TESTS_MISMATCH issue (b/356132258).

Bug: b/352190336
Test: manual test on devices missing essential impl files
Change-Id: I563e8a24b913d47bd5df52cc78c18e4cf72f848d