Fixed: Don't double delete headless edtior client
When the watch face editor is used with a headless client, that
gets explicitly deleted when the editor closes. We don't need the
call to this@WatchFaceImpl.onDestroy() in that scenario because it
would delete the engine twice which is unexpected.

We also don't need to call this@WatchFaceImpl.onDestroy() for non
headless calls unlike in the previous patch.

Test: Tried this with vinson tests broken by the previous patch
Bug: 318768566
Change-Id: I5dd124bb8191ce64a2d88da31315f0cf614c9f3f