Add Glance API for generated previews
This change adds GlanceAppWidget.providePreview as an entrypoint for
users to provide a Composable that is used when creating a preview for
their widget. GlanceAppWidget.composeForPreview is added as an analog to
compose/runComposition in order to run the widget in preview mode and
produce a RemoteViews.

This change also adds GlanceAppWidgetManager.setWidgetPreview as a
convenience API to run the widget in preview mode and set the previews
with AppWidgetManager.setWidgetPreview.

In order to unit test SizeMode.Responsive compositions, the unit test
code is updated to use a real AppWidgetHostView so that size-mapped
RemoteViews can be supported.

Bug: 308041327
Test: GlanceAppWidgetManagerTest.setWidgetPreview
Test: GlanceAppWidgetTest.composeForPreview*
Relnote: "Add Glance APIs for generated previews:
 GlanceAppWidget.providePreview and
 GlanceAppWidgetManager.setWidgetPreview"
Change-Id: Iced163063939f2c3bc2c12bbcfd6cd45bc33df9b