Attempto to fix flaky testLayerBoundsPositionInRotatedView()
Fixes: 350097851

Even after running testLayerBoundsPositionInRotatedView() 1000 times,
I wasn't able to reproduce the flaky result. I suspect that this
is related to something happening on the screen that is outside
the control of the test.

This CL moves the assertions from the test thread to the UI thread
so that the test can't be interrupted while testing at least.

Test: ran test 2000 times
Change-Id: I0c44271f165da65592c30b8cc8e12f34e9a0785b