recyclerview-selection: Introduce specific SOURCE + TOOL_TYPE handlers
Touchpads report their input devices as TOOL_TYPE_FINGER + SOURCE_MOUSE.
This means when the selection code attempts to route events they degrade
to touch events. This is not how a touchpad should behave. Ultimately it
should behave similarly to a mouse.

This introduces the ability to route events using a more specific
handler based on source + toolType and gracefully degrade to the
toolType if the more specific version can't be found.

Test: ./gradlew recyclerview:recyclerview-selection:test
Bug: b:352475728, b:365503054

Change-Id: Ide7c7225552dafa201046dfc1eb4a5b580d18369