Keep a strong reference to created InputConnection in test
These tests started flaking after we changed how InputConnection references are held in memory. WeakReferenced objects can be aggressively collected by the GC if there happens to be no strong reference to them. This is not an issue in regular apps since the platform always keeps a strong reference to the active `InputConnection`.

Here we add a deliberate strong reference so that test concludes successfully.

Fix: 358333887
Test: PlatformTextInputViewIntegrationTest
Change-Id: Idf5f071513e0b554bd96ad92b9e15a4b7dd1472a