Fix a crash when applying a MaterialShape
Fix an issue when a MaterialShape that is applied to a Button with a
border caused a crash.

Seems like the internal Path inside the shape that is created by the
MaterialShape was reused and rescaled when called again. This created a
very large bitmap that caused a crash.

Also, added demos on how to apply shapes from the MaterialShape as a
component Shape, how to clip a simpler component with it, and how to use
it for drawing on a Canvas.

Bug: 360959985
Test: Manual
Change-Id: I5f443c11aa2ae7f63dac1a861cf02447bcbc26f5