Commonize InvalidationTracker Flow API
This changes bring the extension function invalidationTrackerFlow() to the InvalidationTracker as createFlow() and evolves it to be the first-party API for tracking database invalidation in Kotlin Multiplatform.

The code gen utility function FlowBuilder.createFlow() and the extension function invalidationTrackerFlow() are migrated to use the new API.

All InvalidationTracker.Observer APIs that where commonized are moved to the Android source set, to continue keeping backwards compatibility in Android and no longer provide the callback-based APIs for KMP.

LimitOffsetPagingSource was updated to use the Flow API since Observers are no longer available in common code. The internal API to check for invalidation synchronously was updated to return a boolean if interested tables where invalidated, this is needed to better guaranteed invalidation of page source. Tests are also updated to filter out a named Coroutine, a strategy that is less brittle than filtering runnable based on function names.

The implementation uses a MutableStateFlow of an array of ever incrementing table versions. The actual version value are not persisted and don't even correspond to individual table modification. They can be considered as a counter of times the table has been invalidated since the instantiation of the InvalidationTracker. Observing flows keep a reference to the latest versions and as they receive new versions they check if any of the interested tables have change in count, an indicator that the flow must emit.

Bug: 261205680
Bug: 287085166
Bug: 287366097
Bug: 287477564
Bug: 287517337
Bug: 312434479
Bug: 329291639
Bug: 329315924

Test: Existing
Relnote: "Commonize invalidationTrackerFlow() as a first-party API in InvalidationTracker and make it available for non-Android source sets in KMP projects."
Change-Id: I8fb29221e382b7a78259c47d9d93add4c59584ce