Ink: Add strokes module
A snapshot of internal state from cl/666937093, which is based on
cl/667938488. The only changes not from internal state are the Gradle
files. Now that the strokes module is present, the geometry module has
been updated to include its tests that depend on the brush and strokes
modules.

Note that this encountered issues running `jvmTest` in presubmit, but did not have issues with the same tests being run on an Android emulator. JVM tests have been disabled while this is investigated in b/362697089, but the code still has full test coverage through Android emulator tests. It is believed that the way the linux-x86_64 version of the .so file is being generated may be causing issues specific to non-Android JVM builds.

Test: ./gradlew :ink:ink-strokes:ftlpixel2api26

Relnote: Add Ink Strokes module and update Brush and Geometry modules

Change-Id: I1b1edd5aa945a93494ce3ceb44870fa9669cef93