Define base Node class.
Define base Node class for ASTs in AppSearch that all node types must
derive from. The only thing in common is that all nodes must implement a
getChildren method regardless of whether or not they have child nodes
or not in order to making traversal easier.

Bug: 356661489
Bug: 356662352

Relnote: Add Node interface to AppSearch for defining nodes.
Test: No tests for interface.
Change-Id: If42fba0055d383f208eaa287af3e3b77d1bb06da