Limit UUID support in DAOs to Android platform.
This is not migrated to common, and is resulting invalid codegen in non-Android platforms.
The issue was was reproduced and verified in the KMP test app by defining a database and a dao in the QueryTest.kt of the JVM desktop source set that had a DAO function returning an entity with an UUID column. The fix was verified as the failure to import `convertByteToUUID` function error was resolved, and replaced with a "Cannot figure out how to save this field into database. You can consider adding a type converter for it." error.

Adding a type converter allows for the test to pass successfully.

Bug: 362486575
Test: QueryTest.kt on JVM target
Change-Id: I5c9b8e4303ea337e47f6ef59a12a87347b0590dd