Ink: Make Brush Android extensions public
Adjust them a bit to match API feedback for prior modules. This includes a
collection of extension functions, as well as a BrushUtil object that has
static methods for Java callers. The extension functions are marked
JvmSynthetic in favor of explicit alternatives on BrushUtil, as the
ergonomics of extension functions converted to Java functions are not
always great (in particular, extension functions on companion objects).

Test: ./gradlew :ink:ink-brush:ftlpixel2api26

Relnote: Helpers for Ink Brush to use android.graphics.Color instances

Change-Id: I8b2d49713d517a27ba5870dfc2d4f5b10d69a0ec