Separate route from deeplinks in NavDestination
NavDestination's route is no longer added to the list of deeplinks publicly associated with this NavDestination. This means the NavDestination cannot be deeplinked into unless through a deeplink explicity added to the destination by the user.

Test: ./gradlew navigation:navigation-common:cC
Bug: 360084191
Relnote: "A NavDestination can no longer be deeplinked into except through a deeplink that was explicity added to the destination. This also means that you can only navigate to a destination's route with the navigate overload that takes a string route."
Change-Id: Ie30e6c782fc511f1868dfab0d5ab11eab2bf8c4f