Remove usages of the word 'Cursor' in error messages.
With Room KMP out, some error messages need to be updated to be more KMP relevant (or not Android-specific). This change replaces these usages and updates tests accordingly.

Bug: 334087492
Test: Existing tests.
Relnote: Removing and replacing all warnings and error messages in Room that use the word `Cursor` which is no longer accurate in the KMP version of Room.
Change-Id: Id8cd9a6a58809fd974808605b60b8c7ec5f853e8