Update ProfileInstaller 1.3.1 -> 1.4.0-rc01 in macrobenchmarked libraries
Bug: 364907587
Fixes: 364923721
Test: ABTD

This allows these macrobenchmarks to run benchmarks on API 35, which
requires 1.4.0

To keep the dependencies of the macrobenchmarks themselves as minimal
as possible, updating dependencies of the libraries themselves, as
they're both in alpha.

Once 1.4.0 stable lands, we can update all references throughout
AndroidX.

Also cleaned up redundant dependencies in tv macrobench.

Change-Id: Ie16bab57a25908e4ac365a9626e923d2748c8eb4