Improve resolving cap exchange type
The library previously had strict assumptions on when
we to call TelecomManager#getPhoneAccount based
on a permission grant. This is not correct because
the dialer role also grants access to this API
without that permission grant.

Instead, use a more leanient check, where we catch
the SecurityException instead of trying to infer if
access to the API is granted. This is done because
apps using this API should already be default dialer
anyway.

Bug: 365123897
Test: core-telecom tests
Change-Id: I25accef43b625ad74f506d02f4711d012ea9a55d