Wait for the EVENT_CALL_READY signal before starting cap exchange
For Android U/V devices, there is a race condition between sending
the CAPABILITY_EXCHANGE event and telecom internally setting up the
connection. If telecom has not setup the underlying connection, it
can cause the CAPABILITY_EXCHANGE event to be dropped, which causes
extensions to fail to be setup.

This CL introduces a EVENT_CALL_READY event, which is used as a
signal from the VOIP app to the remote calling surfaces and notifies
these applications when the calling application has finished
setting up the telecom connection to the call. For devices running
U/V, we will now wait 500 ms for this event before sending the
capability exchange event.

This CL also changes the tests to have call setup be async so that
we can reproduce this race condition in testing.

Bug: 364311190
Test: manual app testing and core-telecom testing
Change-Id: I7f11ca7b8e65228ebe0eaf190a13236f3af636df