Optimize watch face start up performance
In addition to the API change to support lazy Icons, this patch
makes ResourceDisplayTextWithIndex lazily evaluate the string
which is a decent load performace win for complex watchfaces and
it delays user Style schema validation until after the first frame has been rendered, to avoid this lower priority task competing for cpu/gpu resources.

Relnote: We've added optional support for constructing UserStyleSettings and UserStyleOptions with lazily evaluated Icons.  This is a performance win because it moves Icon construction off the critical path of watch face initialization.
Test: Manual test with perfetto to confirm that lazy icon init is a perf win
Bug: 365522217
Change-Id: Iaf43d401f17a1cca4d1f3a48e363e14fa19c8bcd