Await when abortCapturesOnStop
When a CaptureSessionState is disconnected and
CameraGraph.Flags.abortCapturesOnStop is true, we'll invoke
abortCaptures on the current session, which also triggers a flush on the
camera device inside the camera framework. Since the `disconnect` call
is done in `async`, there's a possibility the `disconnect` call ends up
aborting captures while a new capture session is configuring, causing
unwanted behaviors.

Await when abortCapturesOnStop to prevent the aforementioned scenario.

Bug: 365171667
Test: CameraPipe unit test.
Change-Id: I846a1e6c08cb9699ad3dd5712b6e47d42c41944b