Utilize the Stable Biometric 1.1.0 Dependency
This utilizes the older 1.1.0 library found
[here](https://developer.android.com/jetpack/androidx/releases/biometric#version_110_3)
for our AndroidX credentials library. This is the stable version of
the biometric library, and will unblock our library to move to beta.

This change means that the [BiometricPromptData] will no longer support
building from CryptoObjects that were built using 'PresentationSession'
nor 'OperationHandle', as this library does not support these two
properties.

However, these libraries can still use some of these properties by
relying on the underlying framework libraries, such as for
'operationHandle'.

It will also no longer be possible to support building back the
CryptoObject in the 'fromBundle', as that was dependent on being able to
form the Jetpack version of a CryptoObject utilizing the new
'operationHandle' based constructor. This has been added as a TODO for
when the biometrics library reaches a new stable version.

This change was validated with QA and does not impact current flows,
though could require changes from any users who have dependencies on
higher CryptoObject flows.

Bug: 367347782
Test: Sample App, QA, Unit Tests Updated
Change-Id: Ic3ad1aaf94809a7c007359cabb5bcfeaa8135411