Set isForeground to true by default
The fallback for active resume on devices between SDK levels Q and S_V2
is activated when CameraPipe thinks it's in foreground. Currently this
value is only changed in the integration layer, meaning vanilla
CameraPipe apps won't have the active resume fallback activated. Set
isForeground to true by default so that we'll have the fallback on
suitable SDK levels by default.

Bug: 344752133
Test: Unit tests
Change-Id: I33ea6495a62839bc355df30651959988d0b92149