Don't do special handling for clearFocus on API 27 and below
Bug: 318968220
Fixes: 350534714

This reverts aosp/2928744 because it is causing problems when
focus first arrives into the ComposeView.

When the IME has a NEXT action from an external View, it
can arrive into the ComposeView's requestFocus() with
isInTouchMode = true. We can't distinguish this behavior
from the clearFocus() behavior that aosp/2928744 was
trying to fix.

Test: new test

Change-Id: I99d9d63c6007823dd7be203ce968496e6c0feb9e