Cache the style, paragraphIntrinsics and paragraph/multiParagraph after each AutoSize layout pass
In some cases, the final font size used for layout is the font size that will be returned by getFontSize(). Currently we will lay out text again, which is unnecessary and worsens performance. Now we cache paragraph or textLayoutResult such that in these cases we don't perform layout again.

Additionally, we set paragraphIntrinsics to null if the layouts aren't the same, as if the layout constraints are changed and we still have AutoSize the old paragraphIntrinsics will be used in layout, so the font size doesn't update.

Test: Layout with different sets of constraints that both overflow and don't overflow. The size of the text now changes; this did not occur previously. Existing tests are unaffected and passing.
Change-Id: I72c899a0de53851f1db2e4568053aea46b310b18