fix NPE when btDevice#getName is null
Some NPE crash reports were generated for backwards comapt calls
when converting a CallAudioState to a CallEndpointCompat.
BluetoothDevice.getName can return a null value which can cause issues
when creating a CallEndpointCompat object.

To prevent this error, the variable that sets the CallEndpointCompat can
be null if set by BluetoothDevice.getName but will have a non-null value
before creating the CallEndpointCompat value

Fixes: 367234694
Test: build
Change-Id: Ie3a40b5ebbbcae2b97cd56e62e5218161c77ab93