Allow bottomButton slot in AlertDialog to be nullable
This change allows the bottomButton slot in AlertDialog to be nullable,
allowing developers to remove it completely when necessary.
This is useful in situations when a button stack is desired within the AlertDialog content.

Bug: 367611697
Test: AlertDialogTest.kt
Relnote: "AlertDialog now supports the ability to remove the default
bottomButton by passing null, providing more flexibility for custom layouts"

Change-Id: I34fa93b55b1041959a6a06f63ee270446992e2be