Add ability to pass additional arguments to benchmark
This CL adds a new instrumentation arg "...output.payload.*" that
is copied into the output BenchmarkData.context.payload.

This data will later be sent to skia from androidx.dev so that we
can pass build time parameters from CI to skia.

As the initial usage of this, we are now passing the APK sha
information to benchmark which will be available in the Skia
dashboard.

For the future, it might make sense to advertise this feature in
case our clients have similar needs.

Bug: 362994770
Test: AndroidTestConfigBuilderTest, ResultWriterTest, PayloadArgumentsTest
Change-Id: I8908ecd7f8fee358b44c15353b42cd3bb40eb689