Fix EditText TextField interop Keyboard flicker
When TextField passes focus to another Composable, it makes a call to InputMethodManager to stop the input method. This call is actually conflated with any other similar calls to the end of frame so that if the next composable that receives focus is another TextField, its startInput takes priority.

However the same logic cannot be applied when focus leaves AndroidComposeView
that hosts the TextField. The logic splits into two in terms of where the focus goes next; another editor or not. If the next View that receives focus is EditText or any other editor, we skip the stopInput or hideKeyboard commands. Otherwise, we run these final commands even if the host view loses focus.

Fix: 318530776
Test: TextInputServiceAndroidCommandDebouncingTest
Change-Id: I6b454cbbb323d5ed5de613251b180a0218b8301d