Use deferred transaction in room-paging.
The root cause of this issue is that when using a Room database that is not an in-memory database (which is what is used in most of our tests), the incorrect usage of a write transaction using a read connection was not caught, as an in-memory database's connection pool only has 1 connection provided, which is a "write" connection by default. Therefore, the incorrect usage of the write transaction using the `immediateTransaction` API was not caught by the tests. This CL adds a test that uses a `databaseBuilder()` to ensure this issue is fixed.

Bug: 368380988
Test: Existing
Change-Id: Ibc96d18400cc366d58a158e6fa5feff9e377402a