Fixed issue with manual clips and elevation
Updated AndroidGraphicsLayer implementation to
record the manual clipping operation within the
GraphicsLayer's displaylist instead of applying
the clip in the destination canvas when the layer
is drawn. This addresses the issue of manual outline
clips being lost when elevation is applied due to
reordering of drawing operations within enable/disableZ
calls.

Updated GraphicsLayer implementation to consume the outline
size in addition to the outline in order for RenderNode#start
to take the maximum of the layer size and outline size to
prevent larger outlines from being clipped unexpectedly

Fixes: 368609970
Test: Added test to AndroidGraphicsLayerTest
Change-Id: I7b8ed70b592079980afba95cb4648aa6f9590ef4