Fix the mistriggering of multiple selection by selection-tracker while swiping on an item of RecyclerView.
The gesture starts at a MotionEvent DOWN for both ItemTouchHelper and the selection-tracker, the selection-tracker will begin the long-press timer as well at this moment.

When the swping gesture is formed, the following MotionEvents (MOVEs and UP) are only dispatched to the ItemTouchHelper by RecyclerView, the selection-tracker does not received any but still counts the long-press timer.

Therefore the mistriggering of the multiple-selection.

The solution is to send a CANCEL MotionEvent to all the other OnItemTouchHelpers when one starts to intercept the following MotionEvent.

Bug: 198972519
Test: Manual test and Unit test

Change-Id: I609b1ddc5b7fa298219ee6f1e9947380e7f71878