[RAW Capture] Expose public APIs for RAW and RAW + JPEG ImageCapture
RelNote: "Add output format APIs for RAW and RAW + JPEG ImageCapture, the device capability check is exposed in ImageCaptureCapabilities#getSupportedOutputFormats. The OUTPUT_FORMAT_RAW is to capture RAW image, which is Adobe DNG format and OUTPUT_FORMAT_RAW_JPEG is to simultaneously capture RAW and JPEG images. The new ImageCapture#takePicture API is used for simultaneous image capture, it needs to provide two OutputFileOptions, the first one is for RAW image and the second one is for JPEG image."

Bug: b/364948957
Test: test on real devices
Change-Id: Ib0f3d6195b59efd878bf6d4b4c8f14d92bb94e6e