Ink: Update from internal state
Update based on internal cl/678322844 which has a diffbase of
cl/677988280. This includes some recent implementation improvements to
include in the first alpha release. This also includes a build.gradle
change to depend on the graphics-core library at its 1.0.0 release
rather than at tip of tree, in response to an error code in the tooling.

This modifies some screenshot tests to use a different set of input points than before, to work around the issue tracked by b/369408056.

Test: ./gradlew :ink:ink-authoring:ftlpixel2api33

Relnote: Ink library improvements
Change-Id: I9bea8b688a95ac6956156d39a04eb719ca92764a